Add support to "append" and "add" to linked listo
Code still needs to be made more readable. I also don't fully understand the
"<" as I had to change its direction for it to work.

I think the code written can still be improved so it's more readable and makes
more sense. I had to rewrite it as the first version didn't work.

I like the progress...